+++ tyrant.h
@@ -24,6 +24,10 @@ enum Fix
     barrier_each_turn,
     dont_evade_mimic_selection,
     leech_increase_max_hp,
+    counter_without_damage,
+    subdue_before_attack,
+    corrosive_protect_armor,
+    poison_after_attacked,
     num_fixes
 };